+++ example/breadcrumb.dart
@@ -1,0 +1,11 @@
+import 'package:meta/meta.dart';
+
+class Breadcrumb {
+  final String description;
+
+  final String text;
+
+  final String url;
+
+  Breadcrumb({this.description, @required this.text, @required this.url});
+}