+++ example/views/home_index.html.g.dart
@@ -5,11 +5,9 @@ import 'layout.html.g.dart';
 class home_index_html extends View {
   String render(HttpRequest request) {
     final out = StringBuffer();
-    // ignore: unused_element
-    String _$e(v) => const HtmlEscape().convert(v.toString());
     out.write('<p>\n');
     out.write('    <a href="');
-    out.write(_$e(SiteLinks.productsIndex));
+    out.write(htmlEscape.convert('${SiteLinks.productsIndex}'));
     out.write('">View our cool products</a>\n');
     out.write('</p>\n');
     final layout = layout_html();