+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+- Small changes in the location of files in the `example` folder
+- Added `example/views/breadcrumbs.html`
+
 ## 0.1.1
 
 - Added build transformer `EdartBuilder` (`edart_builder.dart`)